Added ROS test GetPositionAndSliceAreEqual and fix where was not.
